+++ public/service/buffer_test.go
@@ -86,7 +86,7 @@ func TestStreamMemoryBuffer(t *testing.T) {
 	for ; i < total; i++ {
 		msgBytes := make([][]byte, 1)
 		msgBytes[0] = make([]byte, int(incr))
-		msgBytes[0][0] = byte(i)
+		msgBytes[0][0] = i
 
 		select {
 		// Send to buffer
@@ -107,7 +107,7 @@ func TestStreamMemoryBuffer(t *testing.T) {
 		var outTr types.Transaction
 		select {
 		case outTr = <-b.TransactionChan():
-			assert.Equal(t, i, uint8(outTr.Payload.Get(0).Get()[0]))
+			assert.Equal(t, i, outTr.Payload.Get(0).Get()[0])
 		case <-time.After(time.Second):
 			t.Fatalf("Timed out waiting for unbuffered message %v read", i)
 		}
@@ -123,7 +123,7 @@ func TestStreamMemoryBuffer(t *testing.T) {
 	for i = 0; i <= total; i++ {
 		msgBytes := make([][]byte, 1)
 		msgBytes[0] = make([]byte, int(incr))
-		msgBytes[0][0] = byte(i)
+		msgBytes[0][0] = i
 
 		select {
 		case tChan <- types.NewTransaction(message.New(msgBytes), resChan):
@@ -164,7 +164,7 @@ func TestStreamMemoryBuffer(t *testing.T) {
 	// Extract last message
 	select {
 	case outTr = <-b.TransactionChan():
-		assert.Equal(t, uint8(0), uint8(outTr.Payload.Get(0).Get()[0]))
+		assert.Equal(t, byte(0), outTr.Payload.Get(0).Get()[0])
 		outTr.ResponseChan <- response.NewAck()
 	case <-time.After(time.Second):
 		t.Fatalf("Timed out waiting for final buffered message read")
@@ -182,7 +182,7 @@ func TestStreamMemoryBuffer(t *testing.T) {
 	for i = 1; i <= total; i++ {
 		select {
 		case outTr = <-b.TransactionChan():
-			assert.Equal(t, i, uint8(outTr.Payload.Get(0).Get()[0]))
+			assert.Equal(t, i, outTr.Payload.Get(0).Get()[0])
 		case <-time.After(time.Second):
 			t.Fatalf("Timed out waiting for buffered message %v read", i)
 		}
@@ -229,7 +229,7 @@ func TestStreamBufferClosing(t *testing.T) {
 	for i = 0; i < total; i++ {
 		msgBytes := make([][]byte, 1)
 		msgBytes[0] = make([]byte, int(incr))
-		msgBytes[0][0] = byte(i)
+		msgBytes[0][0] = i
 
 		select {
 		case tChan <- types.NewTransaction(message.New(msgBytes), resChan):
@@ -251,7 +251,7 @@ func TestStreamBufferClosing(t *testing.T) {
 	for i = 0; i < total; i++ {
 		select {
 		case val := <-b.TransactionChan():
-			assert.Equal(t, i, uint8(val.Payload.Get(0).Get()[0]))
+			assert.Equal(t, i, val.Payload.Get(0).Get()[0])
 			val.ResponseChan <- response.NewAck()
 		case <-time.After(time.Second):
 			t.Fatalf("Timed out waiting for final buffered message read")
